AJAX Version
This is the latest version of the app, which uses AJAX to retrieve results from each category. The older version is live at essentialrecipeguide.com. I'm having an issue making the the items from the list of each category clickable and display a larger view of the recipe while using AJAX.